\name{mbest-package}
\alias{mbest-package}
\alias{mbest}
\docType{package}
\title{
  Fast Moment-Based Estimation for Hierarchical Models
}
\description{
  This package implements moment-based fitting methods for
  hierarchical models described in the paper
  "Fast Moment-Based Estimation for Hierarchical Models"
  (Perry, 2015).
}
\details{
    \tabular{ll}{
    Package: \tab mbest\cr
    Type: \tab Package\cr
    Version: \tab 0.3\cr
    Date: \tab 2015-04-20\cr
    License: \tab Apache License, Version 2.0\cr
    }
    
    Basic usage is to call \code{\link{mhglm}}.
}

\author{Patrick O. Perry}

\keyword{package}
\seealso{
    \code{\link{mhglm}},
    \code{\link{fixef.mhglm}},
    \code{\link{ranef.mhglm}},
    \code{\link{VarCorr.mhglm}},
    \code{\link{predict.mhglm}}
}
