\name{mbest-package}
\alias{mbest-package}
\alias{mbest}
\docType{package}
\title{
\packageTitle{mbest}
}
\description{
\packageDescription{mbest}
}
\details{
\packageDESCRIPTION{mbest}
\packageIndices{mbest}
    
Basic usage is to call \code{\link{mhglm}}.
}
\references{
P. O. Perry (2017)
"Fast moment-based estimation for hierarchical models."

N. Zhang, K. Schmaus, and P. O. Perry (2018)
"Fitting deeply-nested hierarchical models to a large book review dataset using moment-based estimators."
}

\keyword{package}
\seealso{
    \code{\link{mhglm}},
    \code{\link{fixef.mhglm}},
    \code{\link{ranef.mhglm}},
    \code{\link{VarCorr.mhglm}},
    \code{\link{predict.mhglm}}.
}
