% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data for testing \code{mhglm_ml}.}
\format{An object of class \code{data.frame} with 3000 rows and 5 columns.}
\usage{
simdata
}
\description{
Simulated data for testing \code{mhglm_ml}.
}
\keyword{datasets}
